Fix assistant stats fallback and renderer split
